+++ src/main/java/org/choon/careerbee/domain/company/service/RecruitmentSyncService.java
@@ -1,0 +1,8 @@
+package org.choon.careerbee.domain.company.service;
+
+import org.choon.careerbee.domain.company.dto.response.SaraminRecruitingResp;
+
+public interface RecruitmentSyncService {
+
+    void persistNewRecruitmentsAndNotify(SaraminRecruitingResp apiResp, boolean isOpenRecruitment);
+}
